Ref #451 - Terminate startup if workers fail to start
If the whois or HTTP workers fail to connect to the database
or preloader on startup, they will send a termination signal
to the main process, terminating IRRd. This prevents starting
up in a broken state.

This does not change behaviour afterwards: if the connection
is lost later, the workers will fail the current query after one
reconnection attempt. They will again attempt reconnection
on future queries. This allows IRRd to recover from a
PostgreSQL restart.