Improve CI performance (#704)
This allows many more CI jobs run parallel. There is some cost in duplicate dependency installs, but as we get free parallel jobs it doesn't really have much impact. We could also fix that by having a prebuild step to build a cache for each Python version, but not really worth it.