Increase journal insertion speed
As the journal table is locked per #685 anyways,
we can guarantee that max(_new_serials_per_source)
is the same as running MAX(serial_nrtm) per source
after the first insert.

The MAX(serial_nrtm) query is actually really slow,
so avoiding this has significant benefits, and
reduces the cost of locking the table.