Fix file descriptor issues in PyPy. #848 for additional context.
This is the proper fix for #827, as that issue was not completely fixed with that change. The issue seems to be caused by the socket.fromfd() call in https://pagure.io/python-daemon/blob/main/f/daemon/daemon.py#_765, which checks whether the process runs under superserver when detach_process=None. We skip this check now by always setting it to True.

The apparent side effect seems to cause confusion in file descriptor handling, where the FD that used to be the log file becomes switched around with others. E.g. in the preloader, it could get mixed up with the Redis socket, causing writes to the supposed log file FD to be sent to Redis instead. Checks with strace confirm this, as well as tcpdumps of Redis traffic. The reasons for this issue, or why it only occurs on PyPy, are a mystery.