Add NRTMv4 client for draft-ietf-grow-nrtm-v4-03
This also tunes the scheduling a bit to fit NRTMv4's
one minute fetches.
An appropriate user agent is now set on all HTTP requests.

Note that there is currently a potential race condition
in signature validation: sometimes the validation will
fail due to this, but will recover.
This is to be addressed in the protocol.